requirements.txt
# Core pipeline
faster-whisper>=1.0.0
python-dotenv>=1.0.0
watchdog>=3.0.0
requests>=2.31.0
caldav>=1.3.7

# Transcription & diarization
numpy>=1.24.0
soundfile>=0.12.0
python-speech-features>=0.6
scikit-learn>=1.3.0

# Krisp browser automation (optional — only for auto-download)
playwright>=1.40.0

# Process management
psutil>=5.9.0
setproctitle>=1.3.0
